drag and select issue seems to be resolved
