+++ src/components/StudyPane/Passage/WordBlock.tsx
@@ -5,7 +5,7 @@ import { eventBus } from '@/lib/eventBus';
 import { BoxDisplayStyle, ColorActionType, ColorType } from "@/lib/types";
 import { wrapText, wordsHasSameColor } from "@/lib/utils";
 import EsvPopover from './EsvPopover';
-import { LanguageContext } from '.';
+import { LanguageContext } from './PassageBlock';
 
 type ZoomLevel = {
   [level: number]: { fontSize: string, fontInPx: string, maxWidthPx: number };
